# Optimize SET/INCR*/DECR*/SETRANGE/APPEND by reducing duplicate computation