--- a/src/t_string.c
+++ b/src/t_string.c
@@ -601,7 +601,9 @@ void incrDecrCommand(client *c, long long incr) {
     signalModifiedKey(c,c->db,c->argv[1]);
     notifyKeyspaceEvent(NOTIFY_STRING,"incrby",c->argv[1],c->db->id);
     server.dirty++;
-    addReplyLongLong(c, value);
+    addReplyProto(c,":",1);
+    addReply(c,new);
+    addReplyProto(c,"\r\n",2);
 }
 
 void incrCommand(client *c) {