--- a/src/t_string.c
+++ b/src/t_string.c
@@ -73,8 +73,8 @@ void setGenericCommand(client *c, int flags, robj *key, robj *val, robj *expire,
         if (getGenericCommand(c) == C_ERR) return;
     }
 
-    dictEntry *de = dbFind(c->db, key->ptr);
-    found = (lookupKeyWriteWithDictEntry(c->db,key,de) != NULL);
+    dictEntry *de;
+    found = (lookupKeyWriteWithDictEntry(c->db,key,&de) != NULL);
 
     if ((flags & OBJ_SET_NX && found) ||
         (flags & OBJ_SET_XX && !found))
@@ -433,8 +433,8 @@ void setrangeCommand(client *c) {
         return;
     }
 
-    dictEntry *de = dbFind(c->db, c->argv[1]->ptr);
-    o = lookupKeyWriteWithDictEntry(c->db,c->argv[1],de);
+    dictEntry *de;
+    o = lookupKeyWriteWithDictEntry(c->db,c->argv[1],&de);
     if (o == NULL) {
         /* Return 0 when setting nothing on a non-existing string */
         if (value_len == 0) {
@@ -574,8 +574,8 @@ void msetnxCommand(client *c) {
 void incrDecrCommand(client *c, long long incr) {
     long long value, oldvalue;
     robj *o, *new;
-    dictEntry *de = dbFind(c->db, c->argv[1]->ptr);
-    o = lookupKeyWriteWithDictEntry(c->db,c->argv[1],de);
+    dictEntry *de;
+    o = lookupKeyWriteWithDictEntry(c->db,c->argv[1],&de);
     if (checkType(c,o,OBJ_STRING)) return;
     if (getLongLongFromObjectOrReply(c,o,&value,NULL) != C_OK) return;
 
@@ -640,8 +640,8 @@ void incrbyfloatCommand(client *c) {
     long double incr, value;
     robj *o, *new;
 
-    dictEntry *de = dbFind(c->db, c->argv[1]->ptr);
-    o = lookupKeyWriteWithDictEntry(c->db,c->argv[1],de);
+    dictEntry *de;
+    o = lookupKeyWriteWithDictEntry(c->db,c->argv[1],&de);
     if (checkType(c,o,OBJ_STRING)) return;
     if (getLongDoubleFromObjectOrReply(c,o,&value,NULL) != C_OK ||
         getLongDoubleFromObjectOrReply(c,c->argv[2],&incr,NULL) != C_OK)
@@ -674,8 +674,8 @@ void appendCommand(client *c) {
     size_t totlen;
     robj *o, *append;
 
-    dictEntry *de = dbFind(c->db, c->argv[1]->ptr);
-    o = lookupKeyWriteWithDictEntry(c->db,c->argv[1],de);
+    dictEntry *de;
+    o = lookupKeyWriteWithDictEntry(c->db,c->argv[1],&de);
     if (o == NULL) {
         /* Create the key */
         c->argv[2] = tryObjectEncoding(c->argv[2]);