- Avoid addReplyLongLong (which converts back to string) the value we already have as a robj, by using addReplyProto + addReply
- Avoid doing dbFind Twice for the same dictEntry on INCR*/DECR*/SETRANGE/APPEND commands.
- Avoid multiple sdslen calls with the same input on setrangeCommand and appendCommand
- Introduce setKeyWithDictEntry, which is like setKey(), but accepts an optional dictEntry input: Avoids the second dictFind in SET command

TODO: 
- [x] confirm green CI 
- [ ] confirm green daily
- [x] cover it on benchmark spec.
- [x] Update this PR with benchmark spec automated comment. check https://github.com/redis/redis/pull/13505#issuecomment-2322608048


------------------------------

## Hotspot data that triggered this effort 

Hotspots sample on incrDecr* commands (we can save 3.3% on the 2nd dictFind)

![image](https://github.com/user-attachments/assets/420f572d-9bf1-4d59-a332-49166a6295d9)

Hotspots on set command (we can save 2.1% on the 2nd dictFind)
![image](https://github.com/user-attachments/assets/e2dbdd5e-b306-4abd-b07c-2773f010422c)


# Measuring the improvements 


```
pip3 install redis-benchmarks-specification==0.1.222
redis-benchmarks-spec-client-runner --tests-regexp "memtier_benchmark-1Mkeys-string-append-1-100B|memtier_benchmark-1Mkeys-string-incrby|memtier_benchmark-1Mkeys-string-incrbyfloat|memtier_benchmark-1Mkeys-string-decr|memtier_benchmark-1Mkeys-string-setrange-100B|memtier_benchmark-1Mkeys-load-string-with-10B-values-pipeline-10|memtier_benchmark-1Mkeys-load-string-with-100B-values-pipeline-10" --flushall_on_every_test_start --flushall_on_every_test_end --override-memtier-test-time 60 --cpuset_start_pos 1
```

## Improvements on STRING SET command

Throughput 

Test Name | baseline unstable ( 3fcddfb61f903d7112da186cba8b1c93a99dc87f ) | comparison( 93fae760a0825c4d9c6fb4d0c089ec6f44d9344a ) | % change
-- | -- | -- | --
----------------------------------------------------------------- | -----------: | -----------: | -----------:
memtier_benchmark-1Mkeys-load-string-with-100B-values-pipeline-10 | 578505 | 592740 | 2.46%
memtier_benchmark-1Mkeys-load-string-with-10B-values-pipeline-10 | 644028 | 663759 | 3.06%

Avg. Latency 

Test Name | baseline unstable ( 3fcddfb61f903d7112da186cba8b1c93a99dc87f ) | comparison( 93fae760a0825c4d9c6fb4d0c089ec6f44d9344a ) | % change
-- | -- | -- | --
----------------------------------------------------------------- | -----------: | -----------: | -----------:
memtier_benchmark-1Mkeys-load-string-with-100B-values-pipeline-10 | 3.455 | 3.372 | 2.40%
memtier_benchmark-1Mkeys-load-string-with-10B-values-pipeline-10 | 3.103 | 3.011 | 2.96%


## Improvements on INCRBY command


Baseline
```
root@hpe10:~/redis# taskset -c 1-5 memtier_benchmark --pipeline 10 --test-time 60  --hide-histogram --command='INCRBY __key__ 1'
(...)
[RUN #1 100%,  60 secs]  0 threads:    45616940 ops,  750371 (avg:  760212) ops/sec, 36.42MB/sec (avg: 36.91MB/sec),  2.66 (avg:  2.63) msec latency

4         Threads
50        Connections per thread
60        Seconds


ALL STATS
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Incrbys    760212.36         2.62864         2.55900         5.15100         9.98300     37795.06 
Totals     760212.36         2.62864         2.55900         5.15100         9.98300     37795.06 
```

Comparison

```
root@hpe10:~/redis# taskset -c 1-5 memtier_benchmark --pipeline 10 --test-time 60  --hide-histogram --command='INCRBY __key__ 1'
Writing results to stdout
[RUN #1] Preparing benchmark client...
[RUN #1] Launching threads now...
[RUN #1 100%,  60 secs]  0 threads:    47423740 ops,  789698 (avg:  790317) ops/sec, 37.92MB/sec (avg: 37.95MB/sec),  2.53 (avg:  2.53) msec latency

4         Threads
50        Connections per thread
60        Seconds


ALL STATS
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Incrbys    790316.92         2.52839         2.44700         4.92700         9.91900     38860.38 
Totals     790316.92         2.52839         2.44700         4.92700         9.91900     38860.38 
```