@fcostaoliveira 2024-08-30
### Automated performance analysis summary

This comment was automatically generated given there is performance data available.

Using platform named: intel64-ubuntu22.04-redis-icx1 to do the comparison.

In summary:
- Detected a total of 6 improvements above the improvement water line.

You can check a comparison in detail via the [grafana link](https://benchmarksredisio.grafana.net/d/1fWbtb7nz/experimental-oss-spec-benchmarks/?var-branch=unstable&var-branch=perf.incrby)

### Comparison between unstable and perf.incrby.

Time Period from 5 months ago. (environment used: oss-standalone)

#### Improvements Table

|                                                                                                                       Test Case                                                                                                                       |Baseline unstable (median obs. +- std.dev)|Comparison perf.incrby (median obs. +- std.dev)|% change (higher-better)|           Note            |
|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|------------------------------------------|----------------------------------------------:|------------------------|---------------------------|
|[memtier_benchmark-1Mkeys-load-string-with-100B-values-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-load-string-with-100B-values-pipeline-10.yml)| 636952 +- 1.4% (7 datapoints)            |                                         658638|3.4%                    |IMPROVEMENT                |
|[memtier_benchmark-1Mkeys-load-string-with-10B-values-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-load-string-with-10B-values-pipeline-10.yml)  | 753946 +- 2.0% (7 datapoints)            |                                         786432|4.3%                    |waterline=2.0%. IMPROVEMENT|
|[memtier_benchmark-1Mkeys-string-append-1-100B-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-string-append-1-100B-pipeline-10.yml)                | 767185                                   |                                         792867|3.3%                    |IMPROVEMENT                |
|[memtier_benchmark-1Mkeys-string-incrby-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-string-incrby-pipeline-10.yml)                              | 891079                                   |                                         906416|1.7%                    |IMPROVEMENT                |
|[memtier_benchmark-1Mkeys-string-incrbyfloat-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-string-incrbyfloat-pipeline-10.yml)                    | 449176                                   |                                         457083|1.8%                    |IMPROVEMENT                |
|[memtier_benchmark-1Mkeys-string-setrange-100B-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-string-setrange-100B-pipeline-10.yml)                | 812508                                   |                                         826357|1.7%                    |IMPROVEMENT                |


Improvements test regexp names: memtier_benchmark-1Mkeys-load-string-with-100B-values-pipeline-10|memtier_benchmark-1Mkeys-load-string-with-10B-values-pipeline-10|memtier_benchmark-1Mkeys-string-append-1-100B-pipeline-10|memtier_benchmark-1Mkeys-string-incrby-pipeline-10|memtier_benchmark-1Mkeys-string-incrbyfloat-pipeline-10|memtier_benchmark-1Mkeys-string-setrange-100B-pipeline-10

<details>
  <summary>Full Results table:</summary>

|                                                                                                                       Test Case                                                                                                                       |Baseline unstable (median obs. +- std.dev)|Comparison perf.incrby (median obs. +- std.dev)|% change (higher-better)|           Note            |
|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|------------------------------------------|----------------------------------------------:|------------------------|---------------------------|
|[memtier_benchmark-1Mkeys-load-string-with-100B-values-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-load-string-with-100B-values-pipeline-10.yml)| 636952 +- 1.4% (7 datapoints)            |                                         658638|3.4%                    |IMPROVEMENT                |
|[memtier_benchmark-1Mkeys-load-string-with-10B-values-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-load-string-with-10B-values-pipeline-10.yml)  | 753946 +- 2.0% (7 datapoints)            |                                         786432|4.3%                    |waterline=2.0%. IMPROVEMENT|
|[memtier_benchmark-1Mkeys-string-append-1-100B-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-string-append-1-100B-pipeline-10.yml)                | 767185                                   |                                         792867|3.3%                    |IMPROVEMENT                |
|[memtier_benchmark-1Mkeys-string-incrby-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-string-incrby-pipeline-10.yml)                              | 891079                                   |                                         906416|1.7%                    |IMPROVEMENT                |
|[memtier_benchmark-1Mkeys-string-incrbyfloat-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-string-incrbyfloat-pipeline-10.yml)                    | 449176                                   |                                         457083|1.8%                    |IMPROVEMENT                |
|[memtier_benchmark-1Mkeys-string-setrange-100B-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1Mkeys-string-setrange-100B-pipeline-10.yml)                | 812508                                   |                                         826357|1.7%                    |IMPROVEMENT                |

</details>